requirements.txt
tweepy==3.8.0
python-decouple==3.3
pymongo==3.9.0